+++ _examples/project/pkg/database/database.go
@@ -1,0 +1,9 @@
+package database
+
+type DB struct {
+	/* ... */
+}
+
+func Open(connString string) *DB {
+	return &DB{}
+}